feat: handle exceptions
