+++ src/Exception/Request/ForbiddenRequestException.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types=1);
+
+namespace Atymic\Twitter\Exception\Request;
+
+class ForbiddenRequestException extends TwitterRequestException
+{
+}
