+++ src/Exception/Request/RequestFailureException.php
@@ -1,0 +1,10 @@
+<?php
+declare(strict_types=1);
+
+namespace Atymic\Twitter\Exception\Request;
+
+use Atymic\Twitter\Exception\TwitterException;
+
+class RequestFailureException extends TwitterException
+{
+}
